[pull] master from localstack:master

+++ localstack/services/cloudformation/provider.py
@@ -42,6 +42,7 @@ from localstack.aws.api.cloudformation import (
     GetTemplateOutput,
     GetTemplateSummaryInput,
     GetTemplateSummaryOutput,
+    IncludePropertyValues,
     InsufficientCapabilitiesException,
     InvalidChangeSetStatusException,
     ListChangeSetsOutput,
@@ -95,8 +96,10 @@ from localstack.services.cloudformation.engine.transformers import (
 )
 from localstack.services.cloudformation.stores import (
     cloudformation_stores,
+    find_active_stack_by_name_or_id,
     find_change_set,
     find_stack,
+    find_stack_by_id,
     get_cloudformation_store,
 )
 from localstack.state import StateVisitor
@@ -298,7 +301,10 @@ class CloudformationProvider(CloudformationApi):
         client_request_token: ClientRequestToken = None,
         **kwargs,
     ) -> None:
-        stack = find_stack(context.account_id, context.region, stack_name)
+        stack = find_active_stack_by_name_or_id(context.account_id, context.region, stack_name)
+        if not stack:
+            # aws will silently ignore invalid stack names - we should do the same
+            return
         deployer = template_deployer.TemplateDeployer(context.account_id, context.region, stack)
         deployer.delete_stack()
 
@@ -368,12 +374,17 @@ class CloudformationProvider(CloudformationApi):
 
         deployer = template_deployer.TemplateDeployer(context.account_id, context.region, stack)
         # TODO: there shouldn't be a "new" stack on update
-        new_stack = Stack(context.account_id, context.region, request, template)
+        new_stack = Stack(
+            context.account_id, context.region, request, template, request["TemplateBody"]
+        )
         new_stack.set_resolved_parameters(resolved_parameters)
         stack.set_resolved_parameters(resolved_parameters)
         stack.set_resolved_stack_conditions(resolved_stack_conditions)
         try:
             deployer.update_stack(new_stack)
+        except NoStackUpdates as e:
+            stack.set_stack_status("UPDATE_COMPLETE")
+            raise ValidationError(str(e))
         except Exception as e:
             stack.set_stack_status("UPDATE_FAILED")
             msg = f'Unable to update stack "{stack_name}": {e}'
@@ -678,6 +689,7 @@ class CloudformationProvider(CloudformationApi):
         )
         # only set parameters for the changeset, then switch to stack on execute_change_set
         change_set.set_resolved_parameters(resolved_parameters)
+        change_set.template_body = template_body
 
         # TODO: evaluate conditions
         raw_conditions = transformed_template.get("Conditions", {})
@@ -705,16 +717,16 @@ class CloudformationProvider(CloudformationApi):
         if not changes:
             change_set.metadata["Status"] = "FAILED"
             change_set.metadata["ExecutionStatus"] = "UNAVAILABLE"
-            change_set.metadata[
-                "StatusReason"
-            ] = "The submitted information didn't contain changes. Submit different information to create a change set."
+            change_set.metadata["StatusReason"] = (
+                "The submitted information didn't contain changes. Submit different information to create a change set."
+            )
         else:
-            change_set.metadata[
-                "Status"
-            ] = "CREATE_COMPLETE"  # technically for some time this should first be CREATE_PENDING
-            change_set.metadata[
-                "ExecutionStatus"
-            ] = "AVAILABLE"  # technically for some time this should first be UNAVAILABLE
+            change_set.metadata["Status"] = (
+                "CREATE_COMPLETE"  # technically for some time this should first be CREATE_PENDING
+            )
+            change_set.metadata["ExecutionStatus"] = (
+                "AVAILABLE"  # technically for some time this should first be UNAVAILABLE
+            )
 
         return CreateChangeSetOutput(StackId=change_set.stack_id, Id=change_set.change_set_id)
 
@@ -725,8 +737,10 @@ class CloudformationProvider(CloudformationApi):
         change_set_name: ChangeSetNameOrId,
         stack_name: StackNameOrId = None,
         next_token: NextToken = None,
+        include_property_values: IncludePropertyValues = None,
         **kwargs,
     ) -> DescribeChangeSetOutput:
+        # TODO add support for include_property_values
         # only relevant if change_set_name isn't an ARN
         if not ARN_CHANGESET_REGEX.match(change_set_name):
             if not stack_name:
@@ -874,14 +888,19 @@ class CloudformationProvider(CloudformationApi):
         next_token: NextToken = None,
         **kwargs,
     ) -> DescribeStackEventsOutput:
-        state = get_cloudformation_store(context.account_id, context.region)
+        if stack_name is None:
+            raise ValidationError(
+                "1 validation error detected: Value null at 'stackName' failed to satisfy constraint: Member must not be null"
+            )
 
-        events = []
-        for stack_id, stack in state.stacks.items():
-            if stack_name in [None, stack.stack_name, stack.stack_id]:
-                events.extend(stack.events)
-
-        return DescribeStackEventsOutput(StackEvents=events)
+        stack = find_active_stack_by_name_or_id(context.account_id, context.region, stack_name)
+        if not stack:
+            stack = find_stack_by_id(
+                account_id=context.account_id, region_name=context.region, stack_id=stack_name
+            )
+        if not stack:
+            raise ValidationError(f"Stack [{stack_name}] does not exist")
+        return DescribeStackEventsOutput(StackEvents=stack.events)
 
     @handler("DescribeStackResource")
     def describe_stack_resource(
